fixing the welocome page with responsive
