+++ src/Pages/Bienvenida/Bienvenida.css
@@ -1,0 +1,20 @@
+.custom-shape-divider-bottom-1740484337 {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  overflow: hidden;
+  line-height: 0;
+}
+
+.custom-shape-divider-bottom-1740484337 svg {
+  position: relative;
+  display: block;
+  width: calc(100% + 1.3px);
+  height: 167px;
+  transform: rotateY(180deg);
+}
+
+.custom-shape-divider-bottom-1740484337 .shape-fill {
+  fill: #444e50b3;
+}
